pyproject.toml
[tool.poetry]
name = "nowslides"
version = "0.1.0"
description = ""
authors = ["Maxim Laurijssen <maxim.laurijssen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
Mako = "^1.1.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
